requirements-dev.txt
black==24.8.0
pycodestyle==2.12.1
flake8==7.1.1
pyflakes==3.2.0
tomli==2.0.1
platformdirs==4.3.6
click==8.1.7
packaging==24.1
mccabe==0.7.0
mypy-extensions==1.0.0
pathspec==0.12.1
isort==5.13.2
pluggy==1.5.0
pytest==8.3.3
iniconfig==2.0.0
exceptiongroup==1.2.2
pytest-aiohttp==1.0.5
pytest-asyncio==0.24.0
